+++ src/quantise.c
@@ -119,7 +119,7 @@ long quantise(const float * cb, float vec[], float w[], int k, int m, float *se)
 	e = 0.0;
 	for(i=0; i<k; i++) {
 	    diff = cb[j*k+i]-vec[i];
-	    e += pow(diff*w[i],2.0);
+	    e += powf(diff*w[i],2.0);
 	}
 	if (e < beste) {
 	    beste = e;
@@ -1014,7 +1014,7 @@ void aks_to_M2(
 
       model->A[m] = Am;
   }
-  *snr = 10.0*log10(signal/noise);
+  *snr = 10.0*log10f(signal/noise);
 
   TIMER_SAMPLE_AND_LOG2(tpf, "      rec"); 
 }
@@ -1037,7 +1037,7 @@ int encode_Wo(float Wo)
     float norm;
 
     norm = (Wo - Wo_min)/(Wo_max - Wo_min);
-    index = floor(WO_LEVELS * norm + 0.5);
+    index = floorf(WO_LEVELS * norm + 0.5);
     if (index < 0 ) index = 0;
     if (index > (WO_LEVELS-1)) index = WO_LEVELS-1;
 
@@ -1196,7 +1196,7 @@ float speech_to_uq_lsps(float lsp[],
     */
 
     for(i=0; i<=order; i++)
-	ak[i] *= pow(0.994,(float)i);
+	ak[i] *= powf(0.994,(float)i);
 
     roots = lpc_to_lsp(ak, order, lsp, 5, LSP_DELTA1);
     if (roots != order) {
@@ -1634,7 +1634,7 @@ void locate_lsps_jnd_steps(float lsps[], int order)
     step = 25;
     for(i=0; i<2; i++) {
 	lsp_hz = lsps[i]*4000.0/PI;
-	lsp_hz = floor(lsp_hz/step + 0.5)*step;
+	lsp_hz = floorf(lsp_hz/step + 0.5)*step;
 	lsps[i] = lsp_hz*PI/4000.0;
 	if (i) {
 	    if (lsps[i] == lsps[i-1])
@@ -1648,7 +1648,7 @@ void locate_lsps_jnd_steps(float lsps[], int order)
     step = 50;
     for(i=2; i<4; i++) {
 	lsp_hz = lsps[i]*4000.0/PI;
-	lsp_hz = floor(lsp_hz/step + 0.5)*step;
+	lsp_hz = floorf(lsp_hz/step + 0.5)*step;
 	lsps[i] = lsp_hz*PI/4000.0;
 	if (i) {
 	    if (lsps[i] == lsps[i-1])
@@ -1662,7 +1662,7 @@ void locate_lsps_jnd_steps(float lsps[], int order)
     step = 100;
     for(i=4; i<10; i++) {
 	lsp_hz = lsps[i]*4000.0/PI;
-	lsp_hz = floor(lsp_hz/step + 0.5)*step;
+	lsp_hz = floorf(lsp_hz/step + 0.5)*step;
 	lsps[i] = lsp_hz*PI/4000.0;
 	if (i) {
 	    if (lsps[i] == lsps[i-1])
@@ -1708,9 +1708,9 @@ int encode_energy(float e)
     float e_max = E_MAX_DB;
     float norm;
 
-    e = 10.0*log10(e);
+    e = 10.0*log10f(e);
     norm = (e - e_min)/(e_max - e_min);
-    index = floor(E_LEVELS * norm + 0.5);
+    index = floorf(E_LEVELS * norm + 0.5);
     if (index < 0 ) index = 0;
     if (index > (E_LEVELS-1)) index = E_LEVELS-1;
 
@@ -1736,7 +1736,7 @@ float decode_energy(int index)
 
     step = (e_max - e_min)/E_LEVELS;
     e    = e_min + step*(index);
-    e    = pow(10.0,e/10.0);
+    e    = powf(10.0,e/10.0);
 
     return e;
 }
@@ -1792,11 +1792,11 @@ void compute_weights2(const float *x, const float *xp, float *w, int ndim)
      w[1] *= .3;
   }
   /* Higher weight if pitch is stable */
-  if (fabs(x[0]-xp[0])<.2)
+  if (fabsf(x[0]-xp[0])<.2)
   {
      w[0] *= 2;
      w[1] *= 1.5;
-  } else if (fabs(x[0]-xp[0])>.5) /* Lower if not stable */
+  } else if (fabsf(x[0]-xp[0])>.5) /* Lower if not stable */
   {
      w[0] *= .5;
   }
@@ -1854,8 +1854,8 @@ void quantise_WoE(MODEL *model, float *e, float xq[])
   float Wo_min = TWO_PI/P_MAX;
   float Wo_max = TWO_PI/P_MIN;
 
-  x[0] = log10((model->Wo/PI)*4000.0/50.0)/log10(2);
-  x[1] = 10.0*log10(1e-4 + *e);
+  x[0] = log10f((model->Wo/PI)*4000.0/50.0)/log10f(2);
+  x[1] = 10.0*log10f(1e-4 + *e);
 
   compute_weights2(x, xq, w, ndim);
   for (i=0;i<ndim;i++)
@@ -1874,7 +1874,7 @@ void quantise_WoE(MODEL *model, float *e, float xq[])
     Wo = (2^x)*(PI*50)/4000;
   */
   
-  model->Wo = pow(2.0, xq[0])*(PI*50.0)/4000.0;
+  model->Wo = powf(2.0, xq[0])*(PI*50.0)/4000.0;
 
   /* bit errors can make us go out of range leading to all sorts of
      probs like seg faults */
@@ -1884,7 +1884,7 @@ void quantise_WoE(MODEL *model, float *e, float xq[])
 
   model->L  = PI/model->Wo; /* if we quantise Wo re-compute L */
 
-  *e = pow(10.0, xq[1]/10.0);
+  *e = powf(10.0, xq[1]/10.0);
 }
 
 /*---------------------------------------------------------------------------*\
@@ -1912,8 +1912,8 @@ int encode_WoE(MODEL *model, float e, float xq[])
 
   if (e < 0.0) e = 0;  /* occasional small negative energies due LPC round off I guess */
 
-  x[0] = log10((model->Wo/PI)*4000.0/50.0)/log10(2);
-  x[1] = 10.0*log10(1e-4 + e);
+  x[0] = log10f((model->Wo/PI)*4000.0/50.0)/log10f(2);
+  x[1] = 10.0*log10f(1e-4 + e);
 
   compute_weights2(x, xq, w, ndim);
   for (i=0;i<ndim;i++)
@@ -1959,7 +1959,7 @@ void decode_WoE(MODEL *model, float *e, float xq[], int n1)
   }
 
   //printf("dec: %f %f\n", xq[0], xq[1]);
-  model->Wo = pow(2.0, xq[0])*(PI*50.0)/4000.0;
+  model->Wo = powf(2.0, xq[0])*(PI*50.0)/4000.0;
 
   /* bit errors can make us go out of range leading to all sorts of
      probs like seg faults */
@@ -1969,6 +1969,6 @@ void decode_WoE(MODEL *model, float *e, float xq[], int n1)
 
   model->L  = PI/model->Wo; /* if we quantise Wo re-compute L */
 
-  *e = pow(10.0, xq[1]/10.0);
+  *e = powf(10.0, xq[1]/10.0);
 }
